+++ markata/Post.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class Post(BaseModel):
+    title: str = None
+    date: date = None